changes on api

+++ app/Http/Controllers/Api/EntertainerController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Models\User;
 use App\Models\Event;
+use App\Models\BookVenue;
 use Illuminate\Http\Request;
 use App\Models\TalentCategory;
 use App\Models\EntertainerDetail;
@@ -55,17 +56,17 @@ class EntertainerController extends Controller
             }
         }
 
-        if(isset($request->time)){
-        for ($i = 0; $i < count($request->time); $i++) {
-            $data = [
-                'entertainer_details_id'  => $entertainer->id,
-                'time'  => $request->time[$i],
-                'price_package'  => $request->price_package[$i],
-            ];
+        if (isset($request->time)) {
+            for ($i = 0; $i < count($request->time); $i++) {
+                $data = [
+                    'entertainer_details_id'  => $entertainer->id,
+                    'time'  => $request->time[$i],
+                    'price_package'  => $request->price_package[$i],
+                ];
 
-            EntertainerPricePackage::create($data);
+                EntertainerPricePackage::create($data);
+            }
         }
-    }
         $data = EntertainerDetail::with('entertainerEventPhotos', 'entertainerPricePackage')->find($entertainer->id);
         return $this->sendSuccess('Talent created Successfully', compact('data'));
     }
@@ -77,7 +78,7 @@ class EntertainerController extends Controller
     public function getSingleEntertainer($id)
     {
         // , 'entertainerDetail.entertainerPricePackage'
-        $data = User::with('entertainerDetail.entertainerEventPhotos','entertainerDetail.entertainerPricePackage','entertainerDetail.talentCategory')->find($id);
+        $data = User::with('entertainerDetail.entertainerEventPhotos', 'entertainerDetail.entertainerPricePackage', 'entertainerDetail.talentCategory')->find($id);
         if ($data == null) {
             return $this->sendError("Record Not Found!");
         }
@@ -88,15 +89,10 @@ class EntertainerController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'location' => 'required',
-            // 'title' => 'required',
-            'image' => 'required',
             'bio' => 'required',
             'category_id' => 'required',
             'price' => 'required',
-            // 'event_photos' => 'required',
             'description' => 'required',
-            // 'time' => 'required',
-            // 'price_package' => 'required',
         ]);
         if ($validator->fails()) {
             return $this->sendError($validator->errors()->first());
@@ -112,8 +108,9 @@ class EntertainerController extends Controller
         }
         $data['user_id'] = auth()->id();
         $entertainer = EntertainerDetail::find($id)->update($data);
-        EntertainerEventPhotos::where('entertainer_details_id', $id)->delete();
+
         if ($request->file('event_photos')) {
+            EntertainerEventPhotos::where('entertainer_details_id', $id)->delete();
             foreach ($request->file('event_photos') as $data) {
                 $image = hexdec(uniqid()) . '.' . strtolower($data->getClientOriginalExtension());
                 $data->move(public_path('images'), $image);
@@ -123,15 +120,17 @@ class EntertainerController extends Controller
                 ]);
             }
         }
-        EntertainerPricePackage::where('entertainer_details_id', $id)->delete();
-        for ($i = 0; $i < count($request->time); $i++) {
-            $data = [
-                'entertainer_details_id'  => $id,
-                'time'  => $request->time[$i],
-                'price_package'  => $request->price_package[$i],
-            ];
+        if (isset($request->time)) {
+            EntertainerPricePackage::where('entertainer_details_id', $id)->delete();
+            for ($i = 0; $i < count($request->time); $i++) {
+                $data = [
+                    'entertainer_details_id'  => $id,
+                    'time'  => $request->time[$i],
+                    'price_package'  => $request->price_package[$i],
+                ];
 
-            EntertainerPricePackage::create($data);
+                EntertainerPricePackage::create($data);
+            }
         }
         $data = EntertainerDetail::with('entertainerEventPhotos', 'entertainerPricePackage')->find($id);
         return $this->sendSuccess('Entertainer updated Successfully', compact('data'));
@@ -166,5 +165,9 @@ class EntertainerController extends Controller
     {
         EntertainerDetail::destroy($id);
         return $this->sendSuccess('Entertainer deleted Successfully');
+    }
+    public function entertainer_reviews($id){
+        $data=EntertainerDetail::with('User','reviews.user')->find($id);
+        return $this->sendSuccess('Entertainer reviews',compact('data'));
     }
 }
